requirements.txt
python-telegram-bot
python-telegram-bot-calendar
python-telegram-bot-pagination
requests
python-dotenv
peewee